refactor: moving to helper
